indent+jerry warning fix
git-svn-id: svn://svn.savannah.nongnu.org/osip/trunk/osip@131 a617929b-4c5a-41d8-b177-2eda7db2ca6c

+++ parser/hdr_wwwauthenticate.c
@@ -32,12 +32,12 @@ www_authenticate_init (www_authenticate_t ** dest)
     return -1;
   (*dest)->auth_type = NULL;
   (*dest)->realm = NULL;
-  (*dest)->domain = NULL;	/* optionnal */
+  (*dest)->domain = NULL;       /* optionnal */
   (*dest)->nonce = NULL;
-  (*dest)->opaque = NULL;	/* optionnal */
-  (*dest)->stale = NULL;	/* optionnal */
-  (*dest)->algorithm = NULL;	/* optionnal */
-  (*dest)->qop_options = NULL;	/* optionnal (contains a list of qop-value) */
+  (*dest)->opaque = NULL;       /* optionnal */
+  (*dest)->stale = NULL;        /* optionnal */
+  (*dest)->algorithm = NULL;    /* optionnal */
+  (*dest)->qop_options = NULL;  /* optionnal (contains a list of qop-value) */
   return 0;
 }
 
@@ -50,6 +50,7 @@ msg_setwww_authenticate (sip_t * sip, char *hvalue)
 {
   www_authenticate_t *www_authenticate;
   int i;
+
   if (sip == NULL || sip->www_authenticates == NULL)
     return -1;
   i = www_authenticate_init (&www_authenticate);
@@ -62,7 +63,6 @@ msg_setwww_authenticate (sip_t * sip, char *hvalue)
       sfree (www_authenticate);
       return -1;
     }
-
 #ifdef USE_TMP_BUFFER
   sip->message_property = 2;
 #endif
@@ -78,10 +78,10 @@ quoted_string_set (char *name, char *str, char **result, char **next)
     if (*str)
       str++;
     else
-      return -1;		/* bad header format */
+      return -1;                /* bad header format */
 
   if (strlen (str) <= strlen (name))
-    return -1;			/* bad header format... */
+    return -1;                  /* bad header format... */
 #ifndef WIN32
   if (strncasecmp (name, str, strlen (name)) == 0)
 #else
@@ -91,49 +91,49 @@ quoted_string_set (char *name, char *str, char **result, char **next)
       char *quote1;
       char *quote2;
 
-      {				/* the oSIP stack can detect improperly some paramater!!    */
-	/* for exemple "noncecount" (found in vovida!) is deetected */
-	/* as "nonce". This get ri of this behavior.                */
-	char *hack = strchr (str, '=');
+      { /* the oSIP stack can detect improperly some paramater!!    */
+        /* for exemple "noncecount" (found in vovida!) is detected  */
+        /* as "nonce". This get rid of this behavior.               */
+        char *hack = strchr (str, '=');
 
-	while (' ' == *(hack - 1))	/* get rid of extra spaces */
-	  hack--;
-	if (hack - str != strlen (name))
-	  {
-	    *next = str;
-	    return 0;
-	  }
+        while (' ' == *(hack - 1))      /* get rid of extra spaces */
+          hack--;
+        if ((size_t) (hack - str) != strlen (name))
+          {
+            *next = str;
+            return 0;
+          }
       }
+
       quote1 = quote_find (str);
       if (quote1 == NULL)
-	return -1;		/* bad header format... */
+        return -1;              /* bad header format... */
       quote2 = quote_find (quote1 + 1);
       if (quote2 == NULL)
-	return -1;		/* bad header format... */
+        return -1;              /* bad header format... */
       if (quote2 - quote1 == 1)
-	{
-	  /* this is a special case! The quote contains nothing! */
-	  /* example:   Digest opaque="",cnonce=""               */
-	  /* in this case, we just forget the parameter... this  */
-	  /* this should prevent from user manipulating empty    */
-	  /* strings */
-	  *next = quote2 + 1;
-	  return 0;
-	}
+        {
+          /* this is a special case! The quote contains nothing! */
+          /* example:   Digest opaque="",cnonce=""               */
+          /* in this case, we just forget the parameter... this  */
+          /* this should prevent from user manipulating empty    */
+          /* strings */
+          *next = quote2 + 1;
+          return 0;
+        }
       *result = (char *) smalloc (quote2 - quote1 + 2);
       if (*result == NULL)
-	return -1;
+        return -1;
       sstrncpy (*result, quote1, quote2 - quote1 + 1);
-      *next = quote2 + 1;	/* next element start here */
+      *next = quote2 + 1;       /* next element start here */
       if (strlen (quote2) <= 6)
-	{
-	  *next = NULL;
-	  return 0;
-	}
-      *next = quote2 + 1;	/* next element start here */
-    }
-  else
-    *next = str;		/* wrong header asked! */
+        {
+          *next = NULL;
+          return 0;
+        }
+      *next = quote2 + 1;       /* next element start here */
+  } else
+    *next = str;                /* wrong header asked! */
   return 0;
 }
 
@@ -146,16 +146,16 @@ token_set (char *name, char *str, char **result, char **next)
 
   beg = strchr (str, '=');
   if (beg == NULL)
-    return -1;			/* bad header format... */
+    return -1;                  /* bad header format... */
 
   if (strlen (str) < 6)
-    return 0;			/* end of header... */
+    return 0;                   /* end of header... */
 
   while ((' ' == *str) || (',' == *str))
     if (*str)
       str++;
     else
-      return -1;		/* bad header format */
+      return -1;                /* bad header format */
 
 #ifndef WIN32
   if (strncasecmp (name, str, strlen (name)) == 0)
@@ -167,24 +167,23 @@ token_set (char *name, char *str, char **result, char **next)
 
       end = strchr (str, ',');
       if (end == NULL)
-	end = str + strlen (str);	/* This is the end of the header */
+        end = str + strlen (str);       /* This is the end of the header */
 
       if (end - beg < 2)
-	return -1;
+        return -1;
       *result = (char *) smalloc (end - beg);
       if (*result == NULL)
-	return -1;
+        return -1;
       sstrncpy (*result, beg + 1, end - beg - 1);
       sclrspace (*result);
       if (strlen (end) <= 6)
-	{
-	  *next = NULL;
-	  return 0;
-	}
+        {
+          *next = NULL;
+          return 0;
+        }
       *next = end + 1;
-    }
-  else
-    *next = str;		/* next element start here */
+  } else
+    *next = str;                /* next element start here */
   return 0;
 }
 
@@ -202,7 +201,7 @@ www_authenticate_parse (www_authenticate_t * wwwa, char *hvalue)
   char *space = NULL;
   char *next = NULL;
 
-  space = strchr (hvalue, ' ');	/* SEARCH FOR SPACE */
+  space = strchr (hvalue, ' '); /* SEARCH FOR SPACE */
   if (space == NULL)
     return -1;
 
@@ -218,99 +217,98 @@ www_authenticate_parse (www_authenticate_t * wwwa, char *hvalue)
       int parse_ok = 0;
 
       if (quoted_string_set ("realm", space, &(wwwa->realm), &next))
-	return -1;
+        return -1;
       if (next == NULL)
-	return 0;		/* end of header detected! */
+        return 0;               /* end of header detected! */
       else if (next != space)
-	{
-	  space = next;
-	  parse_ok++;
-	}
+        {
+          space = next;
+          parse_ok++;
+        }
       if (quoted_string_set ("domain", space, &(wwwa->domain), &next))
-	return -1;
+        return -1;
       if (next == NULL)
-	return 0;		/* end of header detected! */
+        return 0;               /* end of header detected! */
       else if (next != space)
-	{
-	  space = next;
-	  parse_ok++;
-	}
+        {
+          space = next;
+          parse_ok++;
+        }
       if (quoted_string_set ("nonce", space, &(wwwa->nonce), &next))
-	return -1;
+        return -1;
       if (next == NULL)
-	return 0;		/* end of header detected! */
+        return 0;               /* end of header detected! */
       else if (next != space)
-	{
-	  space = next;
-	  parse_ok++;
-	}
+        {
+          space = next;
+          parse_ok++;
+        }
       if (quoted_string_set ("opaque", space, &(wwwa->opaque), &next))
-	return -1;
+        return -1;
       if (next == NULL)
-	return 0;		/* end of header detected! */
+        return 0;               /* end of header detected! */
       else if (next != space)
-	{
-	  space = next;
-	  parse_ok++;
-	}
+        {
+          space = next;
+          parse_ok++;
+        }
       if (token_set ("stale", space, &(wwwa->stale), &next))
-	return -1;
+        return -1;
       if (next == NULL)
-	return 0;		/* end of header detected! */
+        return 0;               /* end of header detected! */
       else if (next != space)
-	{
-	  space = next;
-	  parse_ok++;
-	}
+        {
+          space = next;
+          parse_ok++;
+        }
       if (token_set ("algorithm", space, &(wwwa->algorithm), &next))
-	return -1;
+        return -1;
       if (next == NULL)
-	return 0;		/* end of header detected! */
+        return 0;               /* end of header detected! */
       else if (next != space)
-	{
-	  space = next;
-	  parse_ok++;
-	}
+        {
+          space = next;
+          parse_ok++;
+        }
       if (quoted_string_set ("qop", space, &(wwwa->qop_options), &next))
-	return -1;
+        return -1;
       if (next == NULL)
-	return 0;		/* end of header detected! */
+        return 0;               /* end of header detected! */
       else if (next != space)
-	{
-	  space = next;
-	  parse_ok++;
-	}
+        {
+          space = next;
+          parse_ok++;
+        }
       if (0 == parse_ok)
-	{
-	  char *quote1, *quote2, *tmp;
+        {
+          char *quote1, *quote2, *tmp;
 
-	  /* CAUTION */
-	  /* parameter not understood!!! I'm too lazy to handle IT */
-	  /* let's simply bypass it */
-	  if (strlen (space) < 1)
-	    return 0;
-	  tmp = strchr (space + 1, ',');
-	  if (tmp == NULL)	/* it was the last header */
-	    return 0;
-	  quote1 = quote_find (space);
-	  if ((quote1 != NULL) && (quote1 < tmp))	/* this may be a quoted string! */
-	    {
-	      quote2 = quote_find (quote1 + 1);
-	      if (quote2 == NULL)
-		return -1;	/* bad header format... */
-	      if (tmp < quote2)	/* the comma is inside the quotes! */
-		space = strchr (quote2, ',');
-	      else
-		space = tmp;
-	      if (space == NULL)	/* it was the last header */
-		return 0;
-	    }
-	  else
-	    space = tmp;
-	  /* continue parsing... */
-	}
+          /* CAUTION */
+          /* parameter not understood!!! I'm too lazy to handle IT */
+          /* let's simply bypass it */
+          if (strlen (space) < 1)
+            return 0;
+          tmp = strchr (space + 1, ',');
+          if (tmp == NULL)      /* it was the last header */
+            return 0;
+          quote1 = quote_find (space);
+          if ((quote1 != NULL) && (quote1 < tmp))       /* this may be a quoted string! */
+            {
+              quote2 = quote_find (quote1 + 1);
+              if (quote2 == NULL)
+                return -1;      /* bad header format... */
+              if (tmp < quote2) /* the comma is inside the quotes! */
+                space = strchr (quote2, ',');
+              else
+                space = tmp;
+              if (space == NULL)        /* it was the last header */
+                return 0;
+          } else
+            space = tmp;
+          /* continue parsing... */
+        }
     }
-  return 0;			/* ok */
+  return 0;                     /* ok */
 }
 
 /* returns the www_authenticate header.            */
@@ -323,10 +321,9 @@ msg_getwww_authenticate (sip_t * sip, int pos, www_authenticate_t ** dest)
 
   *dest = NULL;
   if (list_size (sip->www_authenticates) <= pos)
-    return -1;			/* does not exist */
+    return -1;                  /* does not exist */
 
-  www_authenticate = (www_authenticate_t *)
-    list_get (sip->www_authenticates, pos);
+  www_authenticate = (www_authenticate_t *) list_get (sip->www_authenticates, pos);
 
   *dest = www_authenticate;
   return pos;
@@ -340,7 +337,7 @@ www_authenticate_getauth_type (www_authenticate_t * www_authenticate)
 
 void
 www_authenticate_setauth_type (www_authenticate_t * www_authenticate,
-			       char *auth_type)
+                               char *auth_type)
 {
   www_authenticate->auth_type = (char *) auth_type;
 }
@@ -364,8 +361,7 @@ www_authenticate_getdomain (www_authenticate_t * www_authenticate)
 }
 
 void
-www_authenticate_setdomain (www_authenticate_t * www_authenticate,
-			    char *domain)
+www_authenticate_setdomain (www_authenticate_t * www_authenticate, char *domain)
 {
   www_authenticate->domain = (char *) domain;
 }
@@ -401,8 +397,7 @@ www_authenticate_getopaque (www_authenticate_t * www_authenticate)
 }
 
 void
-www_authenticate_setopaque (www_authenticate_t * www_authenticate,
-			    char *opaque)
+www_authenticate_setopaque (www_authenticate_t * www_authenticate, char *opaque)
 {
   www_authenticate->opaque = (char *) opaque;
 }
@@ -415,7 +410,7 @@ www_authenticate_getalgorithm (www_authenticate_t * www_authenticate)
 
 void
 www_authenticate_setalgorithm (www_authenticate_t * www_authenticate,
-			       char *algorithm)
+                               char *algorithm)
 {
   www_authenticate->algorithm = (char *) algorithm;
 }
@@ -428,7 +423,7 @@ www_authenticate_getqop_options (www_authenticate_t * www_authenticate)
 
 void
 www_authenticate_setqop_options (www_authenticate_t * www_authenticate,
-				 char *qop_options)
+                                 char *qop_options)
 {
   www_authenticate->qop_options = (char *) qop_options;
 }
@@ -573,7 +568,7 @@ www_authenticate_clone (www_authenticate_t * wwwa, www_authenticate_t ** dest)
     return -1;
 
   i = www_authenticate_init (&wa);
-  if (i == -1)			/* allocation failed */
+  if (i == -1)                  /* allocation failed */
     return -1;
   wa->auth_type = sgetcopy (wwwa->auth_type);
   wa->realm = sgetcopy (wwwa->realm);
